Added sweet debugging and fixed changing the background colour with the grid.
Hitting debug creates a debug.txt file in the LearnSwift/ directory. It prints
out the grid and colours are indicated by colours (1 = first colour, 2 = second
colour, by order of position on the grid). Maybe in the future I'll add an actual
index of what they map to but generally this is good enough for now.