Attempted to get tests (failed), improved colour picking
1. Adventures in testing part 1
Kind of got tests working, I think I know how to test none-UI things
but I wasn't able to get the app working with it. I'll need to
separate my grid saving colours / actions upon it into its own class
as right now it's tied to the view which I'm unsure of how to test.

Separating the grid from the view is probably a good thing, however,
I did want to be able to test the view itself so I'll have to revist
that.

2. Easier colour picking from previously used colours
Now you can right click on a colour on the grid and that colour will
be selected, this is reflected in the UI.